requirements.txt
pandas==1.3.5
matplotlib==3.5.1
seaborn==0.12.0
numpy==1.21.5
Pillow==9.0.1
ipynb==0.5.1
